busybox update

+++ user/xupnpd/src/luaxcore.cpp
@@ -1,5 +1,5 @@
 /* 
- * Copyright (C) 2011-2013 Anton Burdinuk
+ * Copyright (C) 2011-2015 Anton Burdinuk
  * clark15b@gmail.com
  * https://tsdemuxer.googlecode.com/svn/trunk/xupnpd
  */
@@ -1719,7 +1719,7 @@ static int lua_http_sendurl(lua_State* L)
                     {
                         *p2=0;
                         status=atoi(p);
-                    }  
+                    }
                 }
             }
         }else
@@ -1766,7 +1766,7 @@ static int lua_http_sendurl(lua_State* L)
         return 2;
     }else
         rc=1;
-    
+
     size_t n;
 
     if(extra_headers>0)
@@ -1799,7 +1799,7 @@ static int lua_http_sendurl(lua_State* L)
 
     if(tmp)
         free(tmp);
-    
+
     return 1;
 }
 
@@ -1913,7 +1913,7 @@ static int lua_http_notify(lua_State* L)
 
     fprintf(fp,
         "NOTIFY %s HTTP/1.1\r\nHost: %s\r\nUser-Agent: %s\r\nConnection: close\r\nContent-Type: text/xml\r\nContent-Length: %i\r\n"
-        "NT: upnp:event\r\nNTS: upnp:propchange\r\nSID: uuid:%s\r\nSEQ: %i\r\nCache-Control: no-cache\r\n\r\n",url.urn,url.vhost,core::user_agent,len,sid,seq);
+        "NT: upnp:event\r\nNTS: upnp:propchange\r\nSID: uuid:%s\r\nSEQ: %i\r\nCache-Control: no-cache\r\n\r\n",url.urn,url.vhost,core::user_agent,(int)len,sid,seq);
     fwrite(data,len,1,fp);
     fflush(fp);
 
@@ -1972,7 +1972,7 @@ static int lua_http_download(lua_State* L)
                         url.urn,url.vhost,core::user_agent);
 
                 if(*post_data)
-                    fprintf(fp,"Content-Length: %d\r\n",post_data_size);
+                    fprintf(fp,"Content-Length: %d\r\n",(int)post_data_size);
 
 //                fprintf(fp,"Accept-Charset: utf-8\r\n");
